+++ tests/detection/test_map.py
@@ -266,6 +266,50 @@ def test_empty_metric():
     metric.compute()
 
 
+@pytest.mark.skipif(_pytest_condition, reason="test requires that pycocotools and torchvision=>0.8.0 is installed")
+def test_missing_pred():
+    """One good detection, one false negative.
+
+    Map should be lower than 1. Actually it is 0.5, but the exact value depends on where we are sampling (i.e. recall's
+    values)
+    """
+    gts = [
+        dict(boxes=torch.Tensor([[10, 20, 15, 25]]), labels=torch.IntTensor([0])),
+        dict(boxes=torch.Tensor([[10, 20, 15, 25]]), labels=torch.IntTensor([0])),
+    ]
+    preds = [
+        dict(boxes=torch.Tensor([[10, 20, 15, 25]]), scores=torch.Tensor([0.9]), labels=torch.IntTensor([0])),
+        # Empty prediction
+        dict(boxes=torch.Tensor([]), scores=torch.Tensor([]), labels=torch.IntTensor([])),
+    ]
+    metric = MeanAveragePrecision()
+    metric.update(preds, gts)
+    result = metric.compute()
+    assert result["map"] < 1, "MAP cannot be 1, as there is a missing prediction."
+
+
+@pytest.mark.skipif(_pytest_condition, reason="test requires that pycocotools and torchvision=>0.8.0 is installed")
+def test_missing_gt():
+    """The symmetric case of test_missing_pred.
+
+    One good detection, one false positive. Map should be lower than 1. Actually it is 0.5, but the exact value depends
+    on where we are sampling (i.e. recall's values)
+    """
+    gts = [
+        dict(boxes=torch.Tensor([[10, 20, 15, 25]]), labels=torch.IntTensor([0])),
+        dict(boxes=torch.Tensor([]), labels=torch.IntTensor([])),
+    ]
+    preds = [
+        dict(boxes=torch.Tensor([[10, 20, 15, 25]]), scores=torch.Tensor([0.9]), labels=torch.IntTensor([0])),
+        dict(boxes=torch.Tensor([[10, 20, 15, 25]]), scores=torch.Tensor([0.95]), labels=torch.IntTensor([0])),
+    ]
+
+    metric = MeanAveragePrecision()
+    metric.update(preds, gts)
+    result = metric.compute()
+    assert result["map"] < 1, "MAP cannot be 1, as there is an image with no ground truth, but some predictions."
+
+
 @pytest.mark.skipif(_pytest_condition, reason="test requires that torchvision=>0.8.0 is installed")
 def test_error_on_wrong_input():
     """Test class input validation."""